feat(quality-gate-api): updated default quality gates
Criteria update:
  - basic-coverage: 5 → 6 (`RESPONSE_CODE_COVERAGE` replaced by `POSITIVE_RESPONSE_CODE_COVERAGE`
    + `OPERATION_SUCCESS_COVERAGE`; `NO_UNDOCUMENTED_RESPONSE_CODES` replaced by
    `NO_UNDOCUMENTED_POSITIVE_RESPONSE_CODES`)
  - full-feature: 10 → 14 (all 4 new criteria added)
  - minimal / dry-run: unchanged
