+++ microservices/quality-gate-api/src/main/java/io/github/bbortt/snow/white/microservices/quality/gate/api/service/QualityGateService.java
@@ -6,31 +6,16 @@
 
 package io.github.bbortt.snow.white.microservices.quality.gate.api.service;
 
-import static io.github.bbortt.snow.white.commons.quality.gate.OpenApiCriteria.ERROR_RESPONSE_CODE_COVERAGE;
-import static io.github.bbortt.snow.white.commons.quality.gate.OpenApiCriteria.HTTP_METHOD_COVERAGE;
-import static io.github.bbortt.snow.white.commons.quality.gate.OpenApiCriteria.NO_UNDOCUMENTED_ERROR_RESPONSE_CODES;
-import static io.github.bbortt.snow.white.commons.quality.gate.OpenApiCriteria.NO_UNDOCUMENTED_POSITIVE_RESPONSE_CODES;
-import static io.github.bbortt.snow.white.commons.quality.gate.OpenApiCriteria.NO_UNDOCUMENTED_RESPONSE_CODES;
-import static io.github.bbortt.snow.white.commons.quality.gate.OpenApiCriteria.PARAMETER_COVERAGE;
-import static io.github.bbortt.snow.white.commons.quality.gate.OpenApiCriteria.PATH_COVERAGE;
-import static io.github.bbortt.snow.white.commons.quality.gate.OpenApiCriteria.REQUIRED_ERROR_FIELDS_COVERAGE;
-import static io.github.bbortt.snow.white.commons.quality.gate.OpenApiCriteria.REQUIRED_PARAMETER_COVERAGE;
-import static io.github.bbortt.snow.white.commons.quality.gate.OpenApiCriteria.RESPONSE_CODE_COVERAGE;
 import static java.lang.Boolean.FALSE;
-import static java.lang.Boolean.TRUE;
 import static org.springframework.util.CollectionUtils.isEmpty;
 
-import io.github.bbortt.snow.white.commons.quality.gate.OpenApiCriteria;
 import io.github.bbortt.snow.white.microservices.quality.gate.api.api.rest.mapper.QualityGateConfigurationMapper;
 import io.github.bbortt.snow.white.microservices.quality.gate.api.domain.model.QualityGateConfiguration;
-import io.github.bbortt.snow.white.microservices.quality.gate.api.domain.model.QualityGateOpenApiCoverageMapping;
-import io.github.bbortt.snow.white.microservices.quality.gate.api.domain.repository.OpenApiCoverageConfigurationRepository;
 import io.github.bbortt.snow.white.microservices.quality.gate.api.domain.repository.QualityGateConfigurationRepository;
 import io.github.bbortt.snow.white.microservices.quality.gate.api.service.exception.ConfigurationDoesNotExistException;
 import io.github.bbortt.snow.white.microservices.quality.gate.api.service.exception.ConfigurationNameAlreadyExistsException;
 import io.github.bbortt.snow.white.microservices.quality.gate.api.service.exception.UnmodifiableConfigurationException;
 import java.util.List;
-import java.util.stream.Stream;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.jspecify.annotations.NonNull;
@@ -47,7 +32,7 @@ public class QualityGateService {
 
   private final QualityGateConfigurationMapper qualityGateConfigurationMapper;
 
-  private final OpenApiCoverageConfigurationRepository openApiCoverageConfigurationRepository;
+  private final DefaultOpenApiQualityGates defaultOpenApiQualityGates;
   private final QualityGateConfigurationRepository qualityGateConfigurationRepository;
 
   @Transactional
@@ -147,12 +132,9 @@ public class QualityGateService {
       "Updating Quality-Gate configuration table with default values"
     );
 
-    var qualityGateConfigurations = Stream.of(
-      getBasicCoverage(),
-      getFullFeature(),
-      getMinimal(),
-      getDryRun()
-    )
+    var qualityGateConfigurations = defaultOpenApiQualityGates
+      .getDefaultOpenApiCoverageConfigurations()
+      .stream()
       .map(qualityGateConfiguration ->
         qualityGateConfigurationRepository
           .findByName(qualityGateConfiguration.getName())
@@ -166,101 +148,5 @@ public class QualityGateService {
       .toList();
 
     qualityGateConfigurationRepository.saveAll(qualityGateConfigurations);
-  }
-
-  private QualityGateConfiguration getBasicCoverage() {
-    var qualityGateConfiguration = QualityGateConfiguration.builder()
-      .name("basic-coverage")
-      .description(
-        "A pragmatic balance of common expectations without requiring deep error validation."
-      )
-      .isPredefined(TRUE)
-      .build();
-
-    addAllOpenApiCriteria(
-      qualityGateConfiguration,
-      Stream.of(
-        PATH_COVERAGE,
-        HTTP_METHOD_COVERAGE,
-        RESPONSE_CODE_COVERAGE,
-        REQUIRED_PARAMETER_COVERAGE,
-        NO_UNDOCUMENTED_RESPONSE_CODES
-      )
-    );
-
-    return qualityGateConfiguration;
-  }
-
-  private QualityGateConfiguration getFullFeature() {
-    var qualityGateConfiguration = QualityGateConfiguration.builder()
-      .name("full-feature")
-      .description(
-        "The most complete and strict configuration, useful for production-readiness or auditing."
-      )
-      .isPredefined(TRUE)
-      .build();
-
-    addAllOpenApiCriteria(
-      qualityGateConfiguration,
-      Stream.of(
-        PATH_COVERAGE,
-        HTTP_METHOD_COVERAGE,
-        RESPONSE_CODE_COVERAGE,
-        ERROR_RESPONSE_CODE_COVERAGE,
-        REQUIRED_PARAMETER_COVERAGE,
-        PARAMETER_COVERAGE,
-        REQUIRED_ERROR_FIELDS_COVERAGE,
-        NO_UNDOCUMENTED_RESPONSE_CODES,
-        NO_UNDOCUMENTED_ERROR_RESPONSE_CODES,
-        NO_UNDOCUMENTED_POSITIVE_RESPONSE_CODES
-      )
-    );
-
-    return qualityGateConfiguration;
-  }
-
-  private QualityGateConfiguration getMinimal() {
-    var qualityGateConfiguration = QualityGateConfiguration.builder()
-      .name("minimal")
-      .description(
-        "Just enough to ensure the API is reachable at all expected endpoints."
-      )
-      .isPredefined(TRUE)
-      .build();
-
-    addAllOpenApiCriteria(qualityGateConfiguration, Stream.of(PATH_COVERAGE));
-
-    return qualityGateConfiguration;
-  }
-
-  private static QualityGateConfiguration getDryRun() {
-    return QualityGateConfiguration.builder()
-      .name("dry-run")
-      .description(
-        "Doesn’t enforce any rules, but may be used to generate reports or test tooling."
-      )
-      .isPredefined(TRUE)
-      .build();
-  }
-
-  private void addAllOpenApiCriteria(
-    QualityGateConfiguration qualityGateConfiguration,
-    Stream<OpenApiCriteria> openApiCriteria
-  ) {
-    openApiCriteria
-      .map(openApiCriterion ->
-        openApiCoverageConfigurationRepository
-          .findByName(openApiCriterion.name())
-          .orElseThrow(IllegalStateException::new)
-      )
-      .map(openApiCoverageConfiguration ->
-        QualityGateOpenApiCoverageMapping.builder()
-          .openApiCoverageConfiguration(openApiCoverageConfiguration)
-          .qualityGateConfiguration(qualityGateConfiguration)
-          .build()
-      )
-      .forEach(mapping ->
-        qualityGateConfiguration.getOpenApiCoverageConfigurations().add(mapping)
-      );
   }
 }
